requirements.txt
python-dotenv==1.0.0
streamlit==1.24.0
cohere==4.45
python-dotenv==1.0.0